generate the excel file for tables and generate image links;
