+++ setup.py
@@ -21,7 +21,8 @@ setup(
         "layoutparser@https://paddleocr.bj.bcebos.com/whl/layoutparser-0.0.0-py3-none-any.whl",
         "protocol",
         "sentry-sdk==1.5.8",
-        "boto3==1.34.74"
+        "boto3==1.34.74",
+        "html2excel==0.0.6"
         ],
 
     author="Ranjan Shrestha",
